requirements.txt
Flask==3.1.3
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
python-dotenv==1.2.2
mysql-connector-python==9.6.0
